+++ src/traits/i-access/i-access.ts
@@ -60,7 +60,10 @@ export default abstract class iAccess {
 		$e.on('block.mod.*.disabled.*', (e: ModEvent) => {
 			if (e.value === 'false' || e.type === 'remove') {
 				$a.off({group: 'blockOnDisable'});
-				component.emit('enable');
+
+				if (e.type !== 'remove' || e.reason === 'removeMod') {
+					component.emit('enable');
+				}
 
 			} else {
 				component.emit('disable');
@@ -80,6 +83,10 @@ export default abstract class iAccess {
 		});
 
 		$e.on('block.mod.*.focused.*', (e: ModEvent) => {
+			if (e.type === 'remove' && e.reason !== 'removeMod') {
+				return;
+			}
+
 			component.emit(e.value === 'false' || e.type === 'remove' ? 'blur' : 'focus');
 		});
 	}