+++ CHANGELOG.md
@@ -29,7 +29,8 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 #### :bug: Bug Fix
 
 * Fixed arguments providing for `sync.mod`
-* Fixed `bCheckbox` and `bRadioButton` native support 
+* Fixed `bCheckbox` and `bRadioButton` native support
+* Fixed props with multiple types
 
 ## v3.0.0-beta.207 (2019-11-07)