+++ CHANGELOG.md
@@ -23,6 +23,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 * Renamed `data-block-params` -> `data-root-component-params` (`iStaticPage`)
 * Removed `bCheckboxGroup`
 * Removed `bCheckboxIcon`
+* Removed `bFlagIcon`
 * Moved `bForm.delegateAction` to `bForm.action` as a function
 * Removed `@bindModTo` decorator